+++ lib/api/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "api"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tonic = "0.6.2"
+prost = "0.9.0"
+prost-types = "0.9.0"
+serde = { version = "~1.0", features = ["derive"] }
+serde_json = "~1.0"
+schemars = { version = "0.8.8", features = ["uuid"] }
+uuid = { version = "0.8", features = ["v4", "serde"] }
+
+segment = {path = "../segment"}
