+++ lib/api/src/grpc/mod.rs
@@ -1,0 +1,4 @@
+pub mod conversions;
+pub mod models;
+#[allow(clippy::all)]
+pub mod qdrant;
