+++ lib/api/src/grpc/models.rs
@@ -3,21 +3,6 @@ use serde;
 use serde::{Deserialize, Serialize};
 use std::fmt::Debug;
 
-#[derive(Debug, Deserialize, Serialize, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum ApiStatus {
-    Ok,
-    Error(String),
-}
-
-#[derive(Debug, Serialize, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct ApiResponse<D: Serialize + Debug> {
-    pub result: Option<D>,
-    pub status: ApiStatus,
-    pub time: f64,
-}
-
 #[derive(Serialize, Deserialize)]
 pub struct VersionInfo {
     pub title: String,
@@ -31,6 +16,21 @@ impl Default for VersionInfo {
             version: env!("CARGO_PKG_VERSION").to_string(),
         }
     }
+}
+
+#[derive(Debug, Deserialize, Serialize, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum ApiStatus {
+    Ok,
+    Error(String),
+}
+
+#[derive(Debug, Serialize, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct ApiResponse<D: Serialize + Debug> {
+    pub result: Option<D>,
+    pub status: ApiStatus,
+    pub time: f64,
 }
 
 #[derive(Debug, Deserialize, Serialize, JsonSchema)]
