+++ lib/collection/Cargo.toml
@@ -33,8 +33,10 @@ env_logger = "0.9.0"
 merge = "0.1.0"
 async-trait = "0.1.52"
 arc-swap = "1.5.0"
+tonic = "0.6.2"
 
 segment = {path = "../segment"}
+api = {path = "../api"}
 
 itertools = "0.10"
 indicatif = "0.16.2"
