+++ lib/collection/src/operations/mod.rs
@@ -1,4 +1,5 @@
 pub mod config_diff;
+mod conversions;
 pub mod payload_ops;
 pub mod point_ops;
 pub mod types;
