+++ lib/storage/Cargo.toml
@@ -26,6 +26,8 @@ async-trait = "0.1.52"
 raft = { version = "0.6.0", features = ["prost-codec"], default-features = false, optional = true }
 prost = { version = "0.7", optional = true }
 serde_cbor = { version = "0.11.2", optional = true }
+tonic = "0.6.2"
 
 segment = {path = "../segment"}
 collection = {path = "../collection"}
+api = {path = "../api"}
