+++ lib/storage/src/content_manager/mod.rs
@@ -1,5 +1,6 @@
 mod alias_mapping;
 pub mod collection_meta_ops;
 mod collections_ops;
+pub mod conversions;
 pub mod errors;
 pub mod toc;
