+++ src/actix/helpers.rs
@@ -1,6 +1,6 @@
-use crate::common::models::{ApiResponse, ApiStatus};
 use actix_web::rt::time::Instant;
 use actix_web::{HttpResponse, Responder};
+use api::grpc::models::{ApiResponse, ApiStatus};
 use serde::Serialize;
 use std::fmt::Debug;
 use storage::content_manager::errors::StorageError;
