+++ src/common/mod.rs
@@ -2,6 +2,5 @@
 pub mod collections;
 #[allow(dead_code)] // May contain functions used in different binaries. Not actually dead
 pub mod helpers;
-pub mod models;
 #[allow(dead_code)] // May contain functions used in different binaries. Not actually dead
 pub mod points;
