--- src/tonic/api/common.rs
@@ -1,11 +1,0 @@
-use storage::content_manager::errors::StorageError;
-
-pub fn error_to_status(error: StorageError) -> tonic::Status {
-    let error_code = match &error {
-        StorageError::BadInput { .. } => tonic::Code::InvalidArgument,
-        StorageError::NotFound { .. } => tonic::Code::NotFound,
-        StorageError::ServiceError { .. } => tonic::Code::Internal,
-        StorageError::BadRequest { .. } => tonic::Code::InvalidArgument,
-    };
-    return tonic::Status::new(error_code, format!("{}", error));
-}
