+++ src/tonic/api/mod.rs
@@ -1,4 +1,2 @@
 pub mod collections_api;
-mod common;
-mod conversions;
 pub mod points_api;
